requirements.txt
flask
gpiozero
pyyaml
kafka-python
paramiko
py_eureka_client